[metric-label-support]: Utility to convert value_list_t to metric_t
This is another step in implementing the proposal defined in the
document  "Adding Labels to Collectd Metrics"
 - https://docs.google.com/document/d/173gGP3tUD3yfN2NNHxCv0BsKsacfDlSyoaQIn7MqLtQ/edit?usp=sharing
This commit adds a utility method to convert from the old metric
represetation to a the new datastructures updated as per the document:
"Compare data structures for metric labels (collectd)"
- https://docs.google.com/document/d/1RZ7YAO6lWcDDvevzxFGVDgv4UQpdgWoT7OCP9Vpb2CU/edit?usp=sharing

This method is designed to be called from plugindispatch_values during
the transition period before read plugins are amended to use the
metric_t data structure directly.

The unit test serves as a example usage.

Signed-off-by: Manoj Srivastava <srivasta@google.com>
Change-Id: Iecfaf06355a53c1ad2951775ee3f331a852ef14d